+++ src/save.rs
@@ -18,6 +18,7 @@ pub struct CraftingSave {
     pub creatures: Option<Vec<crate::creature::SavedCreature>>,
     pub behaviors: std::collections::BTreeMap<u32, crate::creature::CreatureBehavior>,
     pub dragons: crate::creature::DragonSave,
+    pub fish: crate::creature::FishSave,
 }
 #[derive(serde::Serialize, serde::Deserialize)]
 struct SaveV2 {